change euler conversion to degrees and attempt to fix telemetry label size
