+++ common/utils.py
@@ -1,6 +1,6 @@
 from dataclasses import asdict, dataclass
 from time import time_ns
-
+import math
 
 def time_ms():
     """Returns the current time in milliseconds."""
@@ -66,6 +66,32 @@ def init_imu_data() -> dict:
     data["game quaternion"] = quaternion
 
     return data
+
+def euler_from_quaternion(x, y, z, w):
+        """
+        Convert a quaternion into euler angles (roll, pitch, yaw)
+        roll is rotation around x in degrees (counterclockwise)
+        pitch is rotation around y in degrees (counterclockwise)
+        yaw is rotation around z in degrees (counterclockwise)
+        """
+        t0 = +2.0 * (w * x + y * z)
+        t1 = +1.0 - 2.0 * (x * x + y * y)
+        roll_x = math.atan2(t0, t1)
+     
+        t2 = +2.0 * (w * y - z * x)
+        t2 = +1.0 if t2 > +1.0 else t2
+        t2 = -1.0 if t2 < -1.0 else t2
+        pitch_y = math.asin(t2)
+     
+        t3 = +2.0 * (w * z + x * y)
+        t4 = +1.0 - 2.0 * (y * y + z * z)
+        yaw_z = math.atan2(t3, t4)
+
+        roll_x = (roll_x * 180.0) / math.pi
+        pitch_y = (pitch_y * 180.0) / math.pi
+        yaw_z = (yaw_z * 180.0) / math.pi
+     
+        return roll_x, pitch_y, yaw_z # in degrees
 
 @dataclass
 class VelocityVector:
